refactor(messages): unify 1:1 + group composer actions (#723)
useConversationComposerActions and useGroupComposerActions were ~90% duplicate.
Extract the shared orchestration — gallery / camera / GIF / contact / voice /
location flows, permission prompts, and the in-flight flags — into a single
useComposerActions driven by an injected ComposerSendStrategy. Each screen's
hook is now a thin wrapper that only provides its strategy: the send primitive
(sendDirectMessage/sendFileMessage vs sendGroupMessage), the optimistic-append,
and (1:1 only) the share-location confirm dialog.

No behaviour change at the screens — they call the same wrapper hooks with the
same params and destructure the same returns.

Closes #723
